requirements.txt
pyttsx3
SpeechRecognition
pyaudio
python-dotenv
requests
geocoder
pandas
xlrd
openpyxl
googletrans==3.1.0a0